Add basic scene
The scene 'graph' (array) now exists.

VAOs and shaders are now in indexed lists.

3D objects have a shader index, so that the shaders can be shared
and reloaded without stale pointer issues.

3D objects have a VAO index so that they don't also have to carry
all the VBO and other info in them, and the VBOs can still be deleted.

Shaders are updated to version 4.3 so that we can use static uniform
locations instead of strings all over the source files.